refactor: ProductSearchDataRes 공용화
