+++ src/main/java/com/example/wanted_preonboarding_challenge_backend_31/web/product/dto/response/ProductSearchRes.java
@@ -1,10 +1,11 @@
 package com.example.wanted_preonboarding_challenge_backend_31.web.product.dto.response;
 
 import com.example.wanted_preonboarding_challenge_backend_31.shared.dto.pagination.PaginationRes;
+import com.example.wanted_preonboarding_challenge_backend_31.shared.dto.product.ProductSearchDataDto;
 import java.util.List;
 
 public record ProductSearchRes(
-        List<ProductSearchDataRes> items,
+        List<ProductSearchDataDto> items,
         PaginationRes pagination
 ) {
 }
